added new analysis strategy added UML diagram added Random data implementation to SearchInterface (for testing) moved main class and added more subject strings

+++ src/com/wb/paliver/search/SearchInterface.java
@@ -18,5 +18,5 @@ public interface SearchInterface {
 	 * @param query - a string representing the desired search query
 	 * @return The number of results pages for the query
 	 */
-	public int getPageCount(String query) throws IOException;
+	public long getPageCount(String query) throws IOException;
 }
